Module google_mirror1::api

Structs

Account
Represents an account passed into the Account Manager on Glass.

AccountInsertCall
Inserts a new account for a user

AccountMethods
A builder providing access to all methods supported on account resources. It is not used directly, but through the Mirror hub.

Attachment
Represents media content, such as a photo, that can be attached to a timeline item.

AttachmentsListResponse
A list of Attachments. This is the response from the server to GET requests on the attachments collection.

AuthToken
There is no detailed description.

Command
A single menu command that is part of a Contact.

Contact
A person or group that can be used as a creator or a contact.

ContactDeleteCall
Deletes a contact.

ContactGetCall
Gets a single contact by ID.

ContactInsertCall
Inserts a new contact.

ContactListCall
Retrieves a list of contacts for the authenticated user.

ContactMethods
A builder providing access to all methods supported on contact resources. It is not used directly, but through the Mirror hub.

ContactPatchCall
Updates a contact in place. This method supports patch semantics.

ContactUpdateCall
Updates a contact in place.

ContactsListResponse
A list of Contacts representing contacts. This is the response from the server to GET requests on the contacts collection.

Location
A geographic location that can be associated with a timeline item.

LocationGetCall
Gets a single location by ID.

LocationListCall
Retrieves a list of locations for the user.

LocationMethods
A builder providing access to all methods supported on location resources. It is not used directly, but through the Mirror hub.

LocationsListResponse
A list of Locations. This is the response from the server to GET requests on the locations collection.

MenuItem
A custom menu item that can be presented to the user by a timeline item.

MenuValue
A single value that is part of a MenuItem.

Mirror
Central instance to access all Mirror related resource activities

Notification
A notification delivered by the API.

NotificationConfig
Controls how notifications for a timeline item are presented to the user.

Setting
A setting for Glass.

SettingGetCall
Gets a single setting by ID.

SettingMethods
A builder providing access to all methods supported on setting resources. It is not used directly, but through the Mirror hub.

Subscription
A subscription to events on a collection.

SubscriptionDeleteCall
Deletes a subscription.

SubscriptionInsertCall
Creates a new subscription.

SubscriptionListCall
Retrieves a list of subscriptions for the authenticated user and service.

SubscriptionMethods
A builder providing access to all methods supported on subscription resources. It is not used directly, but through the Mirror hub.

SubscriptionUpdateCall
Updates an existing subscription in place.

SubscriptionsListResponse
A list of Subscriptions. This is the response from the server to GET requests on the subscription collection.

TimelineAttachmentDeleteCall
Deletes an attachment from a timeline item.

TimelineAttachmentGetCall
Retrieves an attachment on a timeline item by item ID and attachment ID.

TimelineAttachmentInsertCall
Adds a new attachment to a timeline item.

TimelineAttachmentListCall
Returns a list of attachments for a timeline item.

TimelineDeleteCall
Deletes a timeline item.

TimelineGetCall
Gets a single timeline item by ID.

TimelineInsertCall
Inserts a new item into the timeline.

TimelineItem
Each item in the user's timeline is represented as a TimelineItem JSON structure, described below.

TimelineListCall
Retrieves a list of timeline items for the authenticated user.

TimelineListResponse
A list of timeline items. This is the response from the server to GET requests on the timeline collection.

TimelineMethods
A builder providing access to all methods supported on timeline resources. It is not used directly, but through the Mirror hub.

TimelinePatchCall
Updates a timeline item in place. This method supports patch semantics.

TimelineUpdateCall
Updates a timeline item in place.

UserAction
Represents an action taken by the user that triggered a notification.

UserData
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.