Struct google_mirror1::Attachment

pub struct Attachment {
    pub content_url: Option<String>,
    pub content_type: Option<String>,
    pub id: Option<String>,
    pub is_processing_content: Option<bool>,
}

Represents media content, such as a photo, that can be attached to a timeline item.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

attachments insert timeline (response)

attachments get timeline (response)

Fields

content_url

The URL for the content.

content_type

The MIME type of the attachment.

id

The ID of the attachment.

is_processing_content

Indicates that the contentUrl is not available because the attachment content is still being processed. If the caller wishes to retrieve the content, it should try again later.