Struct google_mirror1::MenuValue

pub struct MenuValue {
    pub icon_url: Option<String>,
    pub state: Option<String>,
    pub display_name: Option<String>,
}

A single value that is part of a MenuItem.

This type is not used in any activity, and only used as part of another schema.

Fields

icon_url

URL of an icon to display with the menu item.

state

The state that this value applies to. Allowed values are: - DEFAULT - Default value shown when displayed in the menuItems list. - PENDING - Value shown when the menuItem has been selected by the user but can still be cancelled. - CONFIRMED - Value shown when the menuItem has been selected by the user and can no longer be cancelled.

display_name

The name to display for the menu item. If you specify this property for a built-in menu item, the default contextual voice command for that menu item is not shown.