Struct google_mirror1::MultiPartReader

pub struct MultiPartReader<'a> {
    // some fields omitted
}

Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Methods

impl<'a> MultiPartReader<'a>

fn reserve_exact(&mut self, cap: usize)

Reserve memory for exactly the given amount of parts

fn add_part(&mut self, reader: &'a mut Read, size: u64, mime_type: Mime) -> &mut MultiPartReader<'a>

Add a new part to the queue of parts to be read on the first read call.

Arguments

headers - identifying the body of the part. It's similar to the header in an ordinary single-part call, and should thus contain the same information. reader - a reader providing the part's body size - the amount of bytes provided by the reader. It will be put onto the header as content-size. mime - It will be put onto the content type

fn mime_type(&self) -> Mime

Returns the mime-type representing our multi-part message. Use it with the ContentType header.