Struct google_mirror1::TimelineMethods

pub struct TimelineMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on timeline resources. It is not used directly, but through the Mirror hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_mirror1 as mirror1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mirror1::Mirror;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Mirror::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `attachments_delete(...)`, `attachments_get(...)`, `attachments_insert(...)`, `attachments_list(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.timeline();

Methods

impl<'a, C, A> TimelineMethods<'a, C, A>

fn attachments_list(&self, item_id: &str) -> TimelineAttachmentListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of attachments for a timeline item.

Arguments

itemId - The ID of the timeline item whose attachments should be listed.

fn insert(&self, request: TimelineItem) -> TimelineInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new item into the timeline.

Arguments

request - No description provided.

fn patch(&self, request: TimelineItem, id: &str) -> TimelinePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a timeline item in place. This method supports patch semantics.

Arguments

request - No description provided.

id - The ID of the timeline item.

fn list(&self) -> TimelineListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of timeline items for the authenticated user.

fn attachments_insert(&self, item_id: &str) -> TimelineAttachmentInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a new attachment to a timeline item.

Arguments

itemId - The ID of the timeline item the attachment belongs to.

fn attachments_delete(&self, item_id: &str, attachment_id: &str) -> TimelineAttachmentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an attachment from a timeline item.

Arguments

itemId - The ID of the timeline item the attachment belongs to.

attachmentId - The ID of the attachment.

fn delete(&self, id: &str) -> TimelineDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a timeline item.

Arguments

id - The ID of the timeline item.

fn update(&self, request: TimelineItem, id: &str) -> TimelineUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a timeline item in place.

Arguments

request - No description provided.

id - The ID of the timeline item.

fn attachments_get(&self, item_id: &str, attachment_id: &str) -> TimelineAttachmentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an attachment on a timeline item by item ID and attachment ID.

Arguments

itemId - The ID of the timeline item the attachment belongs to.

attachmentId - The ID of the attachment.

fn get(&self, id: &str) -> TimelineGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a single timeline item by ID.

Arguments

id - The ID of the timeline item.