Struct google_mirror1::UserAction

pub struct UserAction {
    pub type_: Option<String>,
    pub payload: Option<String>,
}

Represents an action taken by the user that triggered a notification.

This type is not used in any activity, and only used as part of another schema.

Fields

type_

The type of action. The value of this can be: - SHARE - the user shared an item. - REPLY - the user replied to an item. - REPLY_ALL - the user replied to all recipients of an item. - CUSTOM - the user selected a custom menu item on the timeline item. - DELETE - the user deleted the item. - PIN - the user pinned the item. - UNPIN - the user unpinned the item. - LAUNCH - the user initiated a voice command. In the future, additional types may be added. UserActions with unrecognized types should be ignored.

payload

An optional payload for the action.

For actions of type CUSTOM, this is the ID of the custom menu item that was selected.