Struct google_mirror1::Contact

pub struct Contact {
    pub kind: Option<String>,
    pub display_name: Option<String>,
    pub accept_types: Option<Vec<String>>,
    pub accept_commands: Option<Vec<Command>>,
    pub priority: Option<u32>,
    pub source: Option<String>,
    pub phone_number: Option<String>,
    pub sharing_features: Option<Vec<String>>,
    pub type_: Option<String>,
    pub image_urls: Option<Vec<String>>,
    pub id: Option<String>,
    pub speakable_name: Option<String>,
}

A person or group that can be used as a creator or a contact.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert contacts (request|response)

delete contacts (none)

list contacts (none)

update contacts (request|response)

patch contacts (request|response)

get contacts (response)

Fields
kind: Option<String>

The type of resource. This is always mirror#contact.

display_name: Option<String>

The name to display for this contact.

accept_types: Option<Vec<String>>

A list of MIME types that a contact supports. The contact will be shown to the user if any of its acceptTypes matches any of the types of the attachments on the item. If no acceptTypes are given, the contact will be shown for all items.

accept_commands: Option<Vec<Command>>

A list of voice menu commands that a contact can handle. Glass shows up to three contacts for each voice menu command. If there are more than that, the three contacts with the highest priority are shown for that particular command.

priority: Option<u32>

Priority for the contact to determine ordering in a list of contacts. Contacts with higher priorities will be shown before ones with lower priorities.

source: Option<String>

The ID of the application that created this contact. This is populated by the API

phone_number: Option<String>

Primary phone number for the contact. This can be a fully-qualified number, with country calling code and area code, or a local number.

sharing_features: Option<Vec<String>>

A list of sharing features that a contact can handle. Allowed values are: - ADD_CAPTION

type_: Option<String>

The type for this contact. This is used for sorting in UIs. Allowed values are: - INDIVIDUAL - Represents a single person. This is the default. - GROUP - Represents more than a single person.

image_urls: Option<Vec<String>>

Set of image URLs to display for a contact. Most contacts will have a single image, but a "group" contact may include up to 8 image URLs and they will be resized and cropped into a mosaic on the client.

id: Option<String>

An ID for this contact. This is generated by the application and is treated as an opaque token.

speakable_name: Option<String>

Name of this contact as it should be pronounced. If this contact's name must be spoken as part of a voice disambiguation menu, this name is used as the expected pronunciation. This is useful for contact names with unpronounceable characters or whose display spelling is otherwise not phonetic.