Struct google_mirror1::Subscription

pub struct Subscription {
    pub kind: Option<String>,
    pub notification: Option<Notification>,
    pub updated: Option<String>,
    pub collection: Option<String>,
    pub verify_token: Option<String>,
    pub user_token: Option<String>,
    pub operation: Option<Vec<String>>,
    pub id: Option<String>,
    pub callback_url: Option<String>,
}

A subscription to events on a collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert subscriptions (request|response)

delete subscriptions (none)

list subscriptions (none)

update subscriptions (request|response)

Fields
kind: Option<String>

The type of resource. This is always mirror#subscription.

notification: Option<Notification>

Container object for notifications. This is not populated in the Subscription resource.

updated: Option<String>

The time at which this subscription was last modified, formatted according to RFC 3339.

collection: Option<String>

The collection to subscribe to. Allowed values are: - timeline - Changes in the timeline including insertion, deletion, and updates. - locations - Location updates. - settings - Settings updates.

verify_token: Option<String>

A secret token sent to the subscriber in notifications so that it can verify that the notification was generated by Google.

user_token: Option<String>

An opaque token sent to the subscriber in notifications so that it can determine the ID of the user.

operation: Option<Vec<String>>

A list of operations that should be subscribed to. An empty list indicates that all operations on the collection should be subscribed to. Allowed values are: - UPDATE - The item has been updated. - INSERT - A new item has been inserted. - DELETE - The item has been deleted. - MENU_ACTION - A custom menu item has been triggered by the user.

id: Option<String>

The ID of the subscription.

callback_url: Option<String>

The URL where notifications should be delivered (must start with https://).